MNT: If an EventDescriptor has a 'name' it should be a string.
The 'name' field in EventDescriptors was introduced late and has always been
optional. This change leaves it optional, but requires that if the 'name' field
is present, that it is a string. We make this assumption in a couple places
downstream so it would be good to enforce it here.