Include foreign key in exception.
At two places in the code, we can raise
``UnresolveableForeignKeyError``. One is when we encounter a datum_id
that we don't recognize and the other is when we encounter a
resource_uid that we don't recognize.

Currently, in the first case, we monkey-patch the relevant key onto the
exception, but in the second case, we don't.

In this PR, the second case includes the key too. And, in both cases, we
avoid monkey-patching and instead pass the relevant key to the
``__init__`` of ``UnresolveableForeignKeyError``. This is motivated by
[examples in the Python documentation](https://docs.python.org/3/tutorial/errors.html#user-defined-exceptions)
showing that it's OK for user-defined exceptions to have custom
signatures, which I was not sure of when I wrote the first verison of
this code.