Better error when packing empty page.
Currently, calling `pack_event_page()` with no arguments hits an
unexpected code path and thus fails like

```
UnboundLocalError: local variable 'event' referenced before assignment
```

This PR causes it to raise with a more describing `ValueError`. It is
not possible to pack an "empty" Event Page because an Event Page must
have a time and a uid.

The situation in the same for Datum Page, which must have a Resource.
That is also addressed here.

Tests included, should be good to merge.