Apply retry logic to handler instantiation.
We currently use a retry loop (with configurable time-spacing
and number of retries) when exchanging a Datum for a data payload.
We should do the same when exchanging a Resource for handler instance
because that step can and often does involve file I/O as well.

This factors out the retry loop into an internal utility function.