Fix sphinx build.
The `master` branch currently fails to build docs, with this warning:

```
Warning, treated as error:
/home/travis/build/bluesky/event-model/docs/source/release-history.rst:20:Inline literal start-string without end-string.
Makefile:20: recipe for target 'html' failed
```

This was introduced in #168 and released in that broken state as 1.15.0.
In #168 the Travis-CI webhook didn't fire, so it was accidentally merged
without tests having been run. The commit tagged 1.15.0 failed here:

https://travis-ci.org/github/bluesky/event-model/jobs/680576809

which I noticed because the new docs were not published.