Include resource, datum in filler_state.
The `filler_state` is a threadsafe way to side-band information into
"coersion" functions that, for example, return the data as a dask array.
We added this so that the functions could have access to `shape` via the
descriptor. In some cases we also need access to the resource to do that
successfully. In the future, when we support variable-length chunks, we
may also want access to the datum.

Prototyped and interactively tested against FXI data in
https://gist.github.com/danielballan/b4fccac29cd77e7bdfd93eba775d9b81